Tag 1.3.63
git-svn-id: https://plugins.svn.wordpress.org/ceceppa-multilingua/tags/1.3.63@847619 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ includes/languages.php
@@ -95,11 +95,19 @@ class CeceppaMLLanguages {
 				'cml_notice_category' => '',
 				'cml_enabled' => 1,
 				'cml_sort_id' => $_POST['sort-id'][$i],
+				'cml_rtl' => @$_POST[ 'rtl' ][ $id ],
+				'cml_date_format' => @$_POST[ 'dformat' ][ $i ],
 				'cml_flag_path' => $flag_path ),
-			    array('%d', '%s', '%s', '%s', '%s', '%s', '%s', '%s', '%d', '%d', '%s'));
+			    array('%d', '%s', '%s', '%s', '%s', '%s', '%s', '%s', '%d', '%d', '%d', '%s', '%s'));
 
 	    if( substr( $lang, 0, 2 ) != "en" )
 	      $d->download_language($wpdb->insert_id, $_POST['language'][$i]);
+          
+        //Aggiorno una nuova colonna alla tabella
+        $sql = sprintf( "ALTER TABLE %s ADD lang_%d bigint(20) NOT NULL DEFAULT 0", CECEPPA_ML_RELATIONS, $wpdb->insert_id );
+        $wpdb->query( $sql );
+
+        cml_table_language_columns();
 	  endif;
 	} else {
 	    $wpdb->update(CECEPPA_ML_TABLE,
@@ -113,9 +121,11 @@ class CeceppaMLLanguages {
 				'cml_notice_category' => '',
 				'cml_enabled' => $_POST['lang-enabled'][$i],
 				'cml_sort_id' => $_POST['sort-id'][$i],
+				'cml_rtl' => @$_POST[ 'rtl' ][ $id ],
+				'cml_date_format' => @$_POST[ 'dformat' ][ $i ],
 				'cml_flag_path' => $flag_path ),
 			    array('id' => $id),
-			    array('%d', '%s', '%s', '%s', '%s', '%s', '%s', '%s', '%d', '%d', '%s'),
+			    array('%d', '%s', '%s', '%s', '%s', '%s', '%s', '%s', '%d', '%d', '%d', '%s', '%s'),
 			    array('%d'));
 	}
 	
@@ -124,11 +134,16 @@ class CeceppaMLLanguages {
     endfor;
 
     //Delete
-    if(!empty($_POST['delete'])) :
+    if( ! empty( $_POST['delete'] ) && intval( $_POST[ "delete" ] ) > 0 ) :
       $delete = intval($_POST['delete']);
 
       $wpdb->query("DELETE FROM " . CECEPPA_ML_TABLE . " WHERE id = " . $delete);
       $wpdb->query(sprintf("DELETE FROM " . CECEPPA_ML_POSTS . " WHERE cml_post_lang_1 = %d OR cml_post_lang_2 = %d", $delete, $delete));
+      
+      //Remove the column from CECEPPA_ML_RELATIONS table
+      $wpdb->query( "ALTER TABLE " . CECEPPA_ML_RELATIONS . " DROP lang_" . $delete );
+      
+      cml_table_language_columns();
     endif;
   }
 }
